graspPlate and releasePlate fixup

+++ precise_driver/include/precise_driver/pflex_device.h
@@ -60,8 +60,8 @@ namespace precise_driver
         int getMode();
         bool setMode(const int& mode);
         std::string command(const std::string& cmd);
-        bool graspPlate(const int& width, const int& speed, const int& force);
-        bool releasePlate(const int& width, const int& speed, const int& force);
+        bool graspPlate(const int& width, const int& speed, const double& force);
+        bool releasePlate(const int& width, const int& speed, const double& force);
 
         bool operable();
         void update_movej();
